main.c; tree.c/h; Makefile; gram.y;
In my branch I have edited:
-- makefile for tree.h/c
-- minor tweaks to tree.h/c
-- MAJOR EDITs to main.c
-- include tree.h in gram.y

in my MAIN.C
I "manually" built a symbol and watch the ST dump in process. It really helped me understand what is going on.
It shows the order of how things get into the symbol table and has a English translation of int (*a[10])(); which was impressive.
I originally was just working on our tree functions, but figured it wouldnt be hard to try this, and it wasn't.