Works all of up to function 5 no exceptions
Got the thing to convert right, and output matches. Also got it to throw error for illegal break.
Function return type is kind of "hokie" used a global variable to get the function return type for the return production. I did not see how to know what function a return call would be in (only one, but how to know, and the type... beyond me) used a simple stack, pushed the name on and then popped off if and when a return was called. Also in case of multiple returns, I just pushed it back on. At the end of a function I pop. 

I hope I didn't break anything, fixing other things.

The Add operator is probably the most robust (most correct), then the assignment.

Comparators could use some troubleshooting. Only made sure the ones for the test file worked so far.