Got it to "run"
I tried the ty_query_function... seem to be an issue with passing it the right type... Don't know if it was something about it and Unary Conversion, or if it was after it got evaluated...

I just ran it with the library to see what happened, it worked for the first "print_ints()" but the derefs in the backend were all TYPTR, which is because I believe the getTypeTag always passes that back.

Running to:
        *pi = j;
	print_ints();
	j = *pi * **ppi;

both lines caused seg faults. L-value * of course needs to be coded.